Convert to a poster with flexdashboard layout.

A flexdashboard poster is an R Markdown document using a flexdashboard layout to arrange text, figures, tables and other content into a typical grid format as used for conference posters. The print version will be rendered when knitting if the format option render_print is set to TRUE in the YAML header or the function call; currently this requires that the root dir be set to the default or a path relative to the Rmd document location, otherwise it will result in an error.

flex_dashboard_poster(
  fig_width = 6,
  fig_height = 6,
  fig_mobile = FALSE,
  dev = "png",
  self_contained = TRUE,
  orientation = "rows",
  df_print = "kable",
  css = NULL,
  keep_md = FALSE,
  theme = "default",
  highlight = "default",
  md_extensions = NULL,
  render_print = FALSE,
  ...
)

Arguments

fig_width

Default width (in inches) for figures

fig_height

Default height (in inches) for figures

fig_mobile

Create an additional rendering of each R graphics figure optimized for rendering on mobile devices oriented in portrait mode. If TRUE, creates a figure which is 3.75 x 4.80 inches wide; if FALSE, create no additional figure for mobile devices; if a numeric vector of length 2, creates a mobile figure with the specified width and height.

dev

Graphics device to use for figure output (defaults to png)

self_contained

Produce a standalone HTML file with no external dependencies, using data: URIs to incorporate the contents of linked scripts, stylesheets, images, and videos. Note that even for self contained documents MathJax is still loaded externally (this is necessary because of its size).

orientation

Determines whether level 2 headings are treated as dashboard rows or dashboard columns.

df_print

Method to be used for printing data frames. Valid values include "default", "kable", "tibble", and "paged". The "default" method uses a corresponding S3 method of print, typically print.data.frame. The "kable" method uses the knitr::kable function. The "tibble" method uses the tibble package to print a summary of the data frame. The "paged" method creates a paginated HTML table (note that this method is only valid for formats that produce HTML). In addition to the named methods you can also pass an arbitrary function to be used for printing data frames. You can disable the df_print behavior entirely by setting the option rmarkdown.df_print to FALSE. See Data frame printing section in bookdown book for examples.

css

One or more css files to include

keep_md

Keep the markdown file generated by knitting. Note that if this is TRUE then clean_supporting will always be FALSE.

theme

Visual theme ("default", "bootstrap", "cerulean", "journal", "flatly", "readable", "spacelab", "united", "cosmo", "lumen", "paper", "sandstone", "simplex", or "yeti"). The "cosmo" theme is used when "default" is specified.

highlight

Syntax highlighting style. Supported styles include "default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn", and "haddock". Pass NULL to prevent syntax highlighting.

md_extensions

Markdown extensions to be added or removed from the default definition or R Markdown. See the rmarkdown_format for additional details.

render_print

Whether to render the print version (defaults to FALSE).

...

Additional arguments passed to flex_dashboard

Value

R Markdown output format to pass to render

Details

Compared to flexdashboard::flex_dashboard, it:

uses level 2 headings as dashboard rows by default, because posters are usually divided vertically in horizontal sections (e.g. title, body, figures), each of which can contain a different number of columns (i.e. level three headings)

prints data frames using kable to show nice simple tables when there is data frame output from the code chunks

produces larger figures

never creates mobile rendering of the R graphics

never uses retina figures